FurnitureHelper-2024-02-01-test-with-unit-tests/preference_lambda/pyproject.toml
[tool.poetry]
name = "preference_lambda"
version = "0.1.0"
description = "Collects and validates user's room details and preferences, interacts with GPT-4 to structure the preferences, and stores them in DynamoDB."
authors = ["Kate Yanchenko <kate@lamoom.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
boto3 = "^1.33.11"
PyJWT = "^2.1.0"

[tool.poetry.group.dev.dependencies]
boto3 = "^1.29.6"
botocore = "^1.32.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"